backend/wordpress-api/requirements.txt
requests==2.32.4
fastapi==0.115.14
uvicorn==0.35.0
SQLAlchemy==2.0.41
PyMySQL==1.1.1
python-multipart==0.0.20
python-dotenv==1.1.1